Make use of docker-compose
Instead of having to build and run all containers manually 
docker-compose is facilitated. Thus giving easier control.

During the build process the versions of the base image and the program 
versions can be passed from the compose file to the Dockerfile. The 
Dockerfiles have defaults set in case the args are not provided.

-